requirements.txt
# Main App & Data
streamlit
pandas

# AI & LLM
google-generativeai
sentence-transformers
scikit-learn
# openai (optional, no longer needed)

# Forecasting & Stats
prophet
scipy

# Sentiment
vaderSentiment

fastapi
uvicorn[standard]
plotly
python-dotenv


fastapi
uvicorn[standard]
streamlit
pandas
requests

scikit-learn
lightgbm
prophet
scipy

google-generativeai
python-dotenv
sentence-transformers
vaderSentiment


numpy